different database stuff

+++ sickbeard/databases/mainDB.py
@@ -17,13 +17,15 @@
 # along with Sick Beard.  If not, see <http://www.gnu.org/licenses/>.
 
 import sickbeard
+import sys
 import os.path
 
 from sickbeard import db, common, helpers, logger
 from sickbeard.providers.generic import GenericProvider
 
 from sickbeard import encodingKludge as ek
-from sickbeard.name_parser.parser import NameParser, InvalidNameException 
+from sickbeard.name_parser.parser import NameParser, InvalidNameException
+
 
 class MainSanityCheck(db.DBSanityCheck):
 
@@ -76,19 +78,23 @@ class MainSanityCheck(db.DBSanityCheck):
 
         for cur_orphan in sqlResults:
             logger.log(u"Orphan episode detected! episode_id: " + str(cur_orphan["episode_id"]) + " showid: " + str(cur_orphan["showid"]), logger.DEBUG)
-            logger.log(u"Deleting orphan episode with episode_id: "+str(cur_orphan["episode_id"]))
+            logger.log(u"Deleting orphan episode with episode_id: " + str(cur_orphan["episode_id"]))
             self.connection.action("DELETE FROM tv_episodes WHERE episode_id = ?", [cur_orphan["episode_id"]])
 
         else:
             logger.log(u"No orphan episode, check passed")
 
+
 def backupDatabase(version):
-    helpers.backupVersionedFile(db.dbFilename(), version)
+    if not helpers.backupVersionedFile(db.dbFilename(), version):
+        logger.log(u"Abort upgrading database")
+        sys.exit(1)
 
 # ======================
 # = Main DB Migrations =
 # ======================
 # Add new migrations at the bottom of the list; subclass the previous migration.
+
 
 class InitialSchema (db.SchemaUpgrade):
     def test(self):
@@ -104,12 +110,14 @@ class InitialSchema (db.SchemaUpgrade):
         for query in queries:
             self.connection.action(query)
 
+
 class AddTvrId (InitialSchema):
     def test(self):
         return self.hasColumn("tv_shows", "tvr_id")
 
     def execute(self):
         self.addColumn("tv_shows", "tvr_id")
+
 
 class AddTvrName (AddTvrId):
     def test(self):
@@ -118,12 +126,14 @@ class AddTvrName (AddTvrId):
     def execute(self):
         self.addColumn("tv_shows", "tvr_name", "TEXT", "")
 
+
 class AddAirdateIndex (AddTvrName):
     def test(self):
         return self.hasTable("idx_tv_episodes_showid_airdate")
 
     def execute(self):
         self.connection.action("CREATE INDEX idx_tv_episodes_showid_airdate ON tv_episodes(showid,airdate);")
+
 
 class NumericProviders (AddAirdateIndex):
     def test(self):
@@ -157,12 +167,14 @@ class NumericProviders (AddAirdateIndex):
             args = [curResult["action"], curResult["date"], curResult["showid"], curResult["season"], curResult["episode"], curResult["quality"], curResult["resource"], provider]
             self.connection.action(sql, args)
 
+
 class NewQualitySettings (NumericProviders):
     def test(self):
         return self.hasTable("db_version")
 
     def execute(self):
 
+        # no db_version yet, backup as version 0
         backupDatabase(0)
 
         # old stuff that's been removed from common but we need it to upgrade
@@ -226,7 +238,7 @@ class NewQualitySettings (NumericProviders):
             self.connection.action("UPDATE tv_episodes SET status = ? WHERE episode_id = ?", [common.Quality.compositeStatus(common.DOWNLOADED, newQuality), curUpdate["episode_id"]])
 
         # if no updates were done then the backup is useless
-        if didUpdate:
+        if not didUpdate:
             os.remove(db.dbFilename(suffix='v0'))
 
 
@@ -246,7 +258,7 @@ class NewQualitySettings (NumericProviders):
             elif int(curUpdate["quality"]) == BEST:
                 newQuality = common.BEST
             else:
-                logger.log(u"Unknown show quality: "+str(curUpdate["quality"]), logger.WARNING)
+                logger.log(u"Unknown show quality: " + str(curUpdate["quality"]), logger.WARNING)
                 newQuality = None
 
             if newQuality:
@@ -282,13 +294,18 @@ class NewQualitySettings (NumericProviders):
         self.connection.action("CREATE TABLE db_version (db_version INTEGER);")
         self.connection.action("INSERT INTO db_version (db_version) VALUES (?)", [1])
 
+
 class DropOldHistoryTable(NewQualitySettings):
     def test(self):
         return self.checkDBVersion() >= 2
 
     def execute(self):
-        self.connection.action("DROP TABLE history_old")
+        backupDatabase(self.checkDBVersion())
+
+        if self.hasTable("history_old"):
+            self.connection.action("DROP TABLE history_old")
         self.incDBVersion()
+
 
 class UpgradeHistoryForGenericProviders(DropOldHistoryTable):
     def test(self):
@@ -306,6 +323,7 @@ class UpgradeHistoryForGenericProviders(DropOldHistoryTable):
 
         self.incDBVersion()
 
+
 class AddAirByDateOption(UpgradeHistoryForGenericProviders):
     def test(self):
         return self.checkDBVersion() >= 4
@@ -314,23 +332,32 @@ class AddAirByDateOption(UpgradeHistoryForGenericProviders):
         self.connection.action("ALTER TABLE tv_shows ADD air_by_date NUMERIC")
         self.incDBVersion()
 
+
 class ChangeSabConfigFromIpToHost(AddAirByDateOption):
     def test(self):
         return self.checkDBVersion() >= 5
-    
+
     def execute(self):
-        sickbeard.SAB_HOST = 'http://' + sickbeard.SAB_HOST + '/sabnzbd/'
+        if sickbeard.SAB_HOST and not sickbeard.SAB_HOST.startswith('http'):
+            sickbeard.SAB_HOST = 'http://' + sickbeard.SAB_HOST + '/'
         self.incDBVersion()
+
 
 class FixSabHostURL(ChangeSabConfigFromIpToHost):
     def test(self):
         return self.checkDBVersion() >= 6
-    
+
     def execute(self):
-        if sickbeard.SAB_HOST.endswith('/sabnzbd/'):
-            sickbeard.SAB_HOST = sickbeard.SAB_HOST.replace('/sabnzbd/','/')
+        if sickbeard.SAB_HOST:
+            if sickbeard.SAB_HOST.endswith('/sabnzbd/'):
+                sickbeard.SAB_HOST = sickbeard.SAB_HOST.replace('/sabnzbd/','/')
+
+            if not sickbeard.SAB_HOST.endswith('/'):
+                sickbeard.SAB_HOST = sickbeard.SAB_HOST + '/'
+
         sickbeard.save_config()
         self.incDBVersion()
+
 
 class AddLang (FixSabHostURL):
     def test(self):
@@ -339,13 +366,14 @@ class AddLang (FixSabHostURL):
     def execute(self):
         self.addColumn("tv_shows", "lang", "TEXT", "en")
 
+
 class PopulateRootDirs (AddLang):
     def test(self):
         return self.checkDBVersion() >= 7
-    
+
     def execute(self):
         dir_results = self.connection.select("SELECT location FROM tv_shows")
-        
+
         dir_counts = {}
         for cur_dir in dir_results:
             cur_root_dir = ek.ek(os.path.dirname, ek.ek(os.path.normpath, cur_dir["location"]))
@@ -353,29 +381,30 @@ class PopulateRootDirs (AddLang):
                 dir_counts[cur_root_dir] = 1
             else:
                 dir_counts[cur_root_dir] += 1
-        
+
         logger.log(u"Dir counts: "+str(dir_counts), logger.DEBUG)
-        
+
         if not dir_counts:
             self.incDBVersion()
             return
-        
+
         default_root_dir = dir_counts.values().index(max(dir_counts.values()))
-        
+
         new_root_dirs = str(default_root_dir)+'|'+'|'.join(dir_counts.keys())
         logger.log(u"Setting ROOT_DIRS to: "+new_root_dirs, logger.DEBUG)
-        
+
         sickbeard.ROOT_DIRS = new_root_dirs
-        
+
         sickbeard.save_config()
-        
+
         self.incDBVersion()
-        
+
+
 class SetNzbTorrentSettings(PopulateRootDirs):
 
     def test(self):
         return self.checkDBVersion() >= 8
-    
+
     def execute(self):
 
         use_torrents = False
@@ -385,43 +414,45 @@ class SetNzbTorrentSettings(PopulateRootDirs):
             if cur_provider.isEnabled():
                 if cur_provider.providerType == GenericProvider.NZB:
                     use_nzbs = True
-                    logger.log(u"Provider "+cur_provider.name+" is enabled, enabling NZBs in the upgrade")
+                    logger.log(u"Provider " + cur_provider.name + " is enabled, enabling NZBs in the upgrade")
                     break
                 elif cur_provider.providerType == GenericProvider.TORRENT:
                     use_torrents = True
-                    logger.log(u"Provider "+cur_provider.name+" is enabled, enabling Torrents in the upgrade")
+                    logger.log(u"Provider " + cur_provider.name + " is enabled, enabling Torrents in the upgrade")
                     break
 
         sickbeard.USE_TORRENTS = use_torrents
         sickbeard.USE_NZBS = use_nzbs
-        
+
         sickbeard.save_config()
-        
+
         self.incDBVersion()
 
+
 class FixAirByDateSetting(SetNzbTorrentSettings):
-    
+
     def test(self):
         return self.checkDBVersion() >= 9
 
     def execute(self):
-        
+
         shows = self.connection.select("SELECT * FROM tv_shows")
-        
+
         for cur_show in shows:
             if cur_show["genre"] and "talk show" in cur_show["genre"].lower():
                 self.connection.action("UPDATE tv_shows SET air_by_date = ? WHERE tvdb_id = ?", [1, cur_show["tvdb_id"]])
-        
+
         self.incDBVersion()
+
 
 class AddSizeAndSceneNameFields(FixAirByDateSetting):
 
     def test(self):
-        return self.checkDBVersion() >= 10
-    
+        return self.hasColumn("tv_episodes", "file_size") and self.hasColumn("tv_episodes", "release_name")
+
     def execute(self):
 
-        backupDatabase(10)
+        backupDatabase(self.checkDBVersion())
 
         if not self.hasColumn("tv_episodes", "file_size"):
             self.addColumn("tv_episodes", "file_size")
@@ -430,12 +461,12 @@ class AddSizeAndSceneNameFields(FixAirByDateSetting):
             self.addColumn("tv_episodes", "release_name", "TEXT", "")
 
         ep_results = self.connection.select("SELECT episode_id, location, file_size FROM tv_episodes")
-        
+
         logger.log(u"Adding file size to all episodes in DB, please be patient")
         for cur_ep in ep_results:
             if not cur_ep["location"]:
                 continue
-            
+
             # if there is no size yet then populate it for us
             if (not cur_ep["file_size"] or not int(cur_ep["file_size"])) and ek.ek(os.path.isfile, cur_ep["location"]):
                 cur_size = ek.ek(os.path.getsize, cur_ep["location"])
@@ -443,19 +474,19 @@ class AddSizeAndSceneNameFields(FixAirByDateSetting):
 
         # check each snatch to see if we can use it to get a release name from
         history_results = self.connection.select("SELECT * FROM history WHERE provider != -1 ORDER BY date ASC")
-        
+
         logger.log(u"Adding release name to all episodes still in history")
         for cur_result in history_results:
             # find the associated download, if there isn't one then ignore it
             download_results = self.connection.select("SELECT resource FROM history WHERE provider = -1 AND showid = ? AND season = ? AND episode = ? AND date > ?",
                                                     [cur_result["showid"], cur_result["season"], cur_result["episode"], cur_result["date"]])
             if not download_results:
-                logger.log(u"Found a snatch in the history for "+cur_result["resource"]+" but couldn't find the associated download, skipping it", logger.DEBUG)
+                logger.log(u"Found a snatch in the history for " + cur_result["resource"]+" but couldn't find the associated download, skipping it", logger.DEBUG)
                 continue
 
             nzb_name = cur_result["resource"]
             file_name = ek.ek(os.path.basename, download_results[0]["resource"])
-            
+
             # take the extension off the filename, it's not needed
             if '.' in file_name:
                 file_name = file_name.rpartition('.')[0]
@@ -464,20 +495,20 @@ class AddSizeAndSceneNameFields(FixAirByDateSetting):
             ep_results = self.connection.select("SELECT episode_id, status FROM tv_episodes WHERE showid = ? AND season = ? AND episode = ? AND location != ''",
                                                 [cur_result["showid"], cur_result["season"], cur_result["episode"]])
             if not ep_results:
-                logger.log(u"The episode "+nzb_name+" was found in history but doesn't exist on disk anymore, skipping", logger.DEBUG)
+                logger.log(u"The episode " + nzb_name + " was found in history but doesn't exist on disk anymore, skipping", logger.DEBUG)
                 continue
 
             # get the status/quality of the existing ep and make sure it's what we expect 
             ep_status, ep_quality = common.Quality.splitCompositeStatus(int(ep_results[0]["status"]))
             if ep_status != common.DOWNLOADED:
                 continue
-            
+
             if ep_quality != int(cur_result["quality"]):
-                continue 
+                continue
 
             # make sure this is actually a real release name and not a season pack or something
             for cur_name in (nzb_name, file_name):
-                logger.log(u"Checking if "+cur_name+" is actually a good release name", logger.DEBUG)
+                logger.log(u"Checking if " + cur_name + " is actually a good release name", logger.DEBUG)
                 try:
                     np = NameParser(False)
                     parse_result = np.parse(cur_name)
@@ -491,45 +522,47 @@ class AddSizeAndSceneNameFields(FixAirByDateSetting):
 
         # check each snatch to see if we can use it to get a release name from
         empty_results = self.connection.select("SELECT episode_id, location FROM tv_episodes WHERE release_name = ''")
-        
+
         logger.log(u"Adding release name to all episodes with obvious scene filenames")
         for cur_result in empty_results:
-            
+
             ep_file_name = ek.ek(os.path.basename, cur_result["location"])
             ep_file_name = os.path.splitext(ep_file_name)[0]
-            
+
             # I only want to find real scene names here so anything with a space in it is out
             if ' ' in ep_file_name:
                 continue
-            
+
             try:
                 np = NameParser(False)
                 parse_result = np.parse(ep_file_name)
             except InvalidNameException:
                 continue
-        
+
             if not parse_result.release_group:
                 continue
-            
-            logger.log(u"Name "+ep_file_name+" gave release group of "+parse_result.release_group+", seems valid", logger.DEBUG)
+
+            logger.log(u"Name " + ep_file_name + " gave release group of " + parse_result.release_group + ", seems valid", logger.DEBUG)
             self.connection.action("UPDATE tv_episodes SET release_name = ? WHERE episode_id = ?", [ep_file_name, cur_result["episode_id"]])
 
         self.incDBVersion()
 
+
 class RenameSeasonFolders(AddSizeAndSceneNameFields):
 
     def test(self):
-        return self.checkDBVersion() >= 11
-    
+        return self.hasColumn("tv_shows", "flatten_folders")
+
     def execute(self):
 
         if not self.hasColumn("tv_shows", "flatten_folders"):
+            backupDatabase(self.checkDBVersion())
             # rename the column
             self.connection.action("ALTER TABLE tv_shows RENAME TO tmp_tv_shows")
             self.connection.action("CREATE TABLE tv_shows (show_id INTEGER PRIMARY KEY, location TEXT, show_name TEXT, tvdb_id NUMERIC, network TEXT, genre TEXT, runtime NUMERIC, quality NUMERIC, airs TEXT, status TEXT, flatten_folders NUMERIC, paused NUMERIC, startyear NUMERIC, tvr_id NUMERIC, tvr_name TEXT, air_by_date NUMERIC, lang TEXT)")
             sql = "INSERT INTO tv_shows(show_id, location, show_name, tvdb_id, network, genre, runtime, quality, airs, status, flatten_folders, paused, startyear, tvr_id, tvr_name, air_by_date, lang) SELECT show_id, location, show_name, tvdb_id, network, genre, runtime, quality, airs, status, seasonfolders, paused, startyear, tvr_id, tvr_name, air_by_date, lang FROM tmp_tv_shows"
             self.connection.action(sql)
-            
+
             # flip the values to be opposite of what they were before
             self.connection.action("UPDATE tv_shows SET flatten_folders = 2 WHERE flatten_folders = 1")
             self.connection.action("UPDATE tv_shows SET flatten_folders = 1 WHERE flatten_folders = 0")
